Player 2 Object created.
Added TODOs to make this better.

Fixed an error with attack 2 animation as the player was able to move while doing the player 2 animation. We dont want that as if we try to implement freeze frame doing the combos will be harder.

Bug: Currently there is a black rectangle being drawn behind the character image drawing while the animation is taking place so when the two players overlap one of the player gets hidden behind the other.

Additionally when we do the attack the player moves ahead but I think we keep that as if a character was attacking it would move forwards anyways so. But thats something to decide upon.

Btw I had to delete the repo and re fork as somehow my repo was 9 commits behind and 5 commits ahead even though we had the same code and i couldnt upload but thats now fixed so my past commits are gone the ones that were on my forked commits.