Added the collisions but there are issues for when i press q and attack it empties the whole hp becuase it subtracts 10 health while the animation is running. but i fixed it for E by adding hitLanded boolean but if you can find a more efficient method feel free to implement it!
Additionally some of the hitboxes and hurtboxes are off for the rectangle as this is a basic verison for it. But if you have time can you fix that and also can you finish the samurai class so i can just take the objects from there instead of character class. I just created a Attack2 get set but its not proper can u fix that idk how you did the other attack do it the same way.